pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "worldgen"
version = "0.1.0"
description = "A simple and highly customizable procedural world generator"
readme = "README.md"
license = "MIT"
authors = ["Dtar380"]
maintainers = ["Dtar380"]

[tool.poetry.dependencies]
python = "^3.13"
numpy = "^2.4.2"
matplotlib = "^3.10.8"
pillow = "^12.1.1"
scipy = "^1.17.1"
